Airflow based pipeline orchastration/requirements.txt
apache-airflow[amazon,async,postgres]==2.8.0
apache-airflow-providers-amazon==8.15.0
boto3==1.34.0
pandas==2.0.0
